Fix PR gdb/15326 infinite recursion printing Fortran types
[PATCH PR/15326] The name of a user-defined data type in Fortran is
stored as "tag_name" in the "main_type" struct.  When printing out
type information via "f_type_print_base" the "show" argument passed
to the recursive call of "f_type_print_base" is correctly set to 0
when following a pointer, however, the check at the beginning of the
function only checks if "type_name" is non-NULL: in the case of a
user-defined data-type the name is stored as "tag_name" with a NULL
"type_name", which causes the check to fail.  The consequence is an
infinite recursion when printing a recursively nested user-defined
type. Fixed by extending the check to also inspect "tag_name".

2013-09-24  Frank Penczek  <frank.penczek@intel.com>
	    Christoph Weinmann  <christoph.t.weinmann@intel.com>

	* f-typeprint.c (f_type_print_base): Include TYPE_TAG_NAME
	in conditional that ends recursive descend.

testsuite/
	* gdb.fortran/type-with-ptr.f90: Add Fortran example
	to test recursion caused by pointers in user-defined types.
	* gdb.fortran/type-with-ptr.exp: Add test case for
	recursion.

Change-Id: I0b7e346562d6264370156469acaf0fcbd47b5f5c
Signed-off-by: Frank Penczek <frank.penczek@intel.com>